requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
twilio==8.10.0
langchain==0.0.350
pinecone-client==3.0.0
python-multipart==0.0.6
pydantic==2.5.2
pytest==7.4.3 